refactor(camera): clean producer/consumer things
