+++ raspberrypi_central/smart-camera/app/service_manager/run_camera.py
@@ -1,9 +1,9 @@
-import io
 import signal
 from typing import Callable
 
 from camera.camera import Camera
 from camera.camera_config import camera_config
+from camera.camera_producer_consumer import FrameProducer, FrameIAConsumer
 from camera.camera_record import CameraRecord
 
 from camera.camera_factory import camera_factory
@@ -12,10 +12,7 @@ from mqtt.mqtt_client import get_mqtt_client
 from service_manager.roi_camera_from_args import roi_camera_from_args
 from service_manager.service_manager import RunService
 
-from queue import Full, Empty
 import multiprocessing as mp
-
-from utils.rate_limit import rate_limited
 
 CAMERA_WIDTH = camera_config.camera_width
 CAMERA_HEIGHT = camera_config.camera_height
@@ -48,6 +45,11 @@ class RunSmartCamera(RunService):
         # we process one frame at the time. If we loose some frames to analyze it is not an issue.
         queue_model = mp.Queue(maxsize=1)
 
+        """
+        Use of event: start/end recording.
+        Use of queue: put the event reference (uuid) to save the video with it as a filename,
+        to be able to link the event motion with the video (and also pictures).
+        """
         camera_record_event = mp.Event()
         camera_record_queue = mp.Queue(maxsize=1)
 
@@ -88,72 +90,8 @@ class RunSmartCamera(RunService):
         new_roi = roi_camera_from_args(data)
         return new_roi != self._camera_analyze_object
 
-    def stop(self, *args) -> None:
-        pass
-
     def __str__(self):
         return 'run-smart-camera'
-
-
-class FrameProducer:
-    def __init__(self, queues, camera_record_event: mp.Event, camera_record_queue: mp.Queue):
-        self.queues = queues
-        self.camera_record_queue = camera_record_queue
-        self.camera_record_event = camera_record_event
-        self.camera_stream = None
-
-    def set_camera_steam(self, camera_stream):
-        self.camera_stream = camera_stream
-
-    def manage_record(self):
-
-        try:
-            record_event_ref = self.camera_record_queue.get_nowait()
-        except Empty:
-            pass
-        else:
-            self.camera_stream.start_recording(record_event_ref)
-
-        if not self.camera_record_event.is_set():
-            self.camera_stream.stop_recording()
-
-    def produce(self, raw_capture):
-        self.manage_record()
-
-        # get bytes from BytesIO, image is bytes here.
-        image_bytes = raw_capture.read()
-
-        for queue in self.queues:
-            try:
-                queue.put(image_bytes, False)
-            except Full:
-                # if the Queue is full, then we ignore frames.
-                pass
-
-
-class FrameIAConsumer:
-    """
-    Take frames from queue and send it to process.
-    """
-    def __init__(self, camera):
-        self.camera = camera
-
-    def process_frame(self, queue):
-        # start mqtt connection
-        self.camera.start()
-
-        # we don't need to be thread safe because we don't use threads.
-        @rate_limited(1, thread_safe=False)
-        def process_frame():
-            try:
-                stream = io.BytesIO(queue.get_nowait())
-            except Empty:
-                pass
-            else:
-                self.camera.process_frame(stream)
-
-        while True:
-            process_frame()
 
 
 def run_smart_camera_factory():
