+++ raspberrypi_central/smart-camera/app/service_manager/service_manager.py
@@ -14,7 +14,3 @@ class RunService(ABC):
     @abstractmethod
     def is_restart_necessary(self, data) -> bool:
         pass
-
-    @abstractmethod
-    def stop(self, *args) -> None:
-        pass
